requirements.txt
fastapi==0.115.5
openai==1.35.9
pydantic==2.8.0
python-dotenv==1.0.1
uvicorn==0.32.1